+++ core-ng/src/main/java/core/framework/internal/web/HTTPHandler.java
@@ -2,6 +2,7 @@ package core.framework.internal.web;
 
 import core.framework.internal.log.ActionLog;
 import core.framework.internal.log.LogManager;
+import core.framework.internal.log.Trace;
 import core.framework.internal.web.bean.RequestBeanReader;
 import core.framework.internal.web.bean.ResponseBeanWriter;
 import core.framework.internal.web.controller.ControllerHolder;
@@ -131,7 +132,8 @@ public class HTTPHandler implements HttpHandler {
         String correlationId = headers.getFirst(HTTPHandler.HEADER_CORRELATION_ID);
         if (correlationId != null) actionLog.correlationIds = List.of(correlationId);
 
-        if ("true".equals(headers.getFirst(HEADER_TRACE))) actionLog.trace = true;
+        String trace = headers.getFirst(HEADER_TRACE);
+        if (trace != null) actionLog.trace = Trace.parse(trace);
 
         actionLog.maxProcessTime(maxProcessTime(headers.getFirst(HTTPHandler.HEADER_TIMEOUT)));
     }