+++ core-ng/src/test/java/core/framework/internal/log/ActionLogTest.java
@@ -35,8 +35,10 @@ class ActionLogTest {
 
     @Test
     void flushTraceLogWithTrace() {
-        log.trace = true;
+        log.trace = Trace.CURRENT;
+        assertThat(log.flushTraceLog()).isTrue();
 
+        log.trace = Trace.CASCADE;
         assertThat(log.flushTraceLog()).isTrue();
     }