+++ core-ng/src/test/java/core/framework/internal/web/HTTPHandlerTest.java
@@ -1,6 +1,7 @@
 package core.framework.internal.web;
 
 import core.framework.internal.log.ActionLog;
+import core.framework.internal.log.Trace;
 import io.undertow.util.HeaderMap;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -26,7 +27,7 @@ class HTTPHandlerTest {
         headers.put(HTTPHandler.HEADER_CLIENT, "client");
         handler.linkContext(actionLog, headers);
 
-        assertThat(actionLog.trace).isTrue();
+        assertThat(actionLog.trace).isEqualTo(Trace.CURRENT);
         assertThat(actionLog.clients).containsExactly("client");
     }